cmd/tailscale/cli: Allow serve remote destination
This commit is a wip commit, that aims to warn user that they won't be able to serve
remote destinations in certain setups.

This commit hasn't fixed the tests and also has a linux case to check in shouldWarnRemoteDestCompatibility
That case needs a new local api to check. Other than that most remote destinations are fine for
services.

updates tailscale/corp/#24783

Signed-off-by: KevinLiang10 <37811973+KevinLiang10@users.noreply.github.com>