ipn/ipnlocal: track whether an auth is in progress
Before we introduced seamless, the "blocked" state was used to track:

* Whether a login was required for connectivity, and therefore we should
  keep the engine deconfigured until that happened
* Whether authentication was in progress

"blocked" would stop authReconfig from running. We want this when a login is
required: if your key has expired we want to deconfigure the engine and keep
it down, so that you don't keep using exit nodes (which won't work because
your key has expired).

Taking the engine down while auth was in progress was undesirable, so we
don't do that with seamless renewal. However, not entering the "blocked"
state meant that when we entered the authenticated state, we could no longer
tell if we had just brought the tunnel up, or had done a reauthentication.
This in turn meant we were not firing "LoginFinished" events on the bus
reliably anymore, making it harder for UI clients to track when
authentication is complete.

Updates tailscale/corp#31476

Updates tailscale/corp#32645

Signed-off-by: James Sanderson <jsanderson@tailscale.com>